requirements.txt
requests
pandas
streamlit
plotly_express
st-pages